Basic outline for the speaking section
All posts go in the _posts folder and we use the category system to
split them up.

if it's a blog post, use category blog
if it's about speaking, using category speaking

All categories are put in the blog index. Will look into changing this
later.

Only speaking category goes in the speaking index.